+++ gpxviewr/gpxviewr/settings.py
@@ -134,6 +134,8 @@ CELERY_BEAT_SCHEDULE = {
     }
 }
 
+VALHALLA_ROUTING_API_HOST = 'http://localhost:8002/route'
+
 # Static files (CSS, JavaScript, Images)
 # https://docs.djangoproject.com/en/4.2/howto/static-files/